feat: Add object/map input support for table assign.
